refactor: rename saleOrder folder
